requirements_web.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
python-dotenv==1.0.0
requests==2.31.0
APScheduler==3.10.4

# Para desenvolvimento
python-dateutil==2.8.2 